utilizar Hibernate

+++ src/main/java/Torneo/Futbol/Modelo/EquipoRequest.java
@@ -5,6 +5,6 @@ public class EquipoRequest {
    public String entrenador;
 public    String logo;
 public     String nacionalidad;
-public     long pais;
-public     long grupo;
+public     int pais;
+public     int grupo;
 }
